Start work on final packages

+++ bootstrap/chroot.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ "$EUID" -eq 0 ]; then
+    echo "Please don't run as root"
+    exit
+fi
+
 export KISS_PROMPT=0
 
 sudo rm -rf "$KISS_ROOT"
@@ -14,7 +19,7 @@ export RANLIB="$KISS_TGT-ranlib"
 kiss b 1-meta
 kiss b 2-meta
 
-sandbox="$(dirname "$0")"/../sandbox.sh
+sandbox="$(dirname "$0")/../sandbox.sh"
 sudo rm -rf chroot
 sudo KISS_REPO=chroot "${sandbox}" /root/stage3.sh
 sudo "$(dirname "$0")/backup.sh"
